+++ src/libtcod/sys_sdl_img_png.c
@@ -39,6 +39,7 @@
 #include <SDL.h>
 
 #include "../vendor/lodepng.h"
+#include "error.h"
 #include "libtcod_int.h"
 bool TCOD_sys_check_png(const char* filename) {
   static uint8_t magic_number[] = {137, 80, 78, 71, 13, 10, 26, 10};
@@ -99,28 +100,15 @@ SDL_Surface* TCOD_sys_read_png(const char* filename) {
   return bitmap;
 }
 
-void TCOD_sys_write_png(SDL_Surface* surf, const char* filename) {
-  unsigned char *image, *dest = malloc(surf->h * surf->w * 3 * sizeof(*dest));
-  int x, y;
-  unsigned char* buf;
-  size_t size;
-  int error;
-  /* SDL uses 32bits format without alpha layer for screen. convert it to 24 bits */
-  image = dest;
-  for (y = 0; y < surf->h; y++) {
-    for (x = 0; x < surf->w; x++) {
-      uint8_t* pixel = (uint8_t*)surf->pixels + y * surf->pitch + x * surf->format->BytesPerPixel;
-      *dest++ = *((pixel) + surf->format->Rshift / 8);
-      *dest++ = *((pixel) + surf->format->Gshift / 8);
-      *dest++ = *((pixel) + surf->format->Bshift / 8);
-    }
+TCOD_Error TCOD_sys_write_png(SDL_Surface* surf, const char* filename) {
+  SDL_Surface* rgba32 = SDL_ConvertSurfaceFormat(surf, SDL_PIXELFORMAT_RGBA32, 0);
+  if (!rgba32) {
+    return TCOD_set_errorvf("SDL: %s", SDL_GetError());
   }
-  error = lodepng_encode_memory(&buf, &size, image, surf->w, surf->h, LCT_RGB, 8);
-  free(image);
-  if (!error) {
-    TCOD_sys_write_file(filename, buf, (uint32_t)size);
-    free(buf);
-  } else {
-    printf("error %u: %s\n", error, lodepng_error_text(error));
+  unsigned int error = lodepng_encode32_file(filename, rgba32->pixels, rgba32->w, rgba32->h);
+  SDL_FreeSurface(rgba32);
+  if (error) {
+    return TCOD_set_errorvf("LodePNG error %u: %s\n", error, lodepng_error_text(error));
   }
+  return TCOD_E_OK;
 }