+++ CHANGELOG.md
@@ -16,6 +16,7 @@ since `1.7.0`.
 - Video driver status info has been added to SDL video driver errors.
 - Console blits now ignore alpha compositing which would result in division by
   zero.
+- Image loading now returns an error code.
 
 ### Fixed
 - `TCOD_console_is_key_pressed` should work even if libtcod events are ignored.